refator: dry eudiw / ewc logic
