+++ webapp/messages/en.json
@@ -605,5 +605,6 @@
 	"This_item_is_yours": "This {item} is yours",
 	"Make_changes": "Make changes",
 	"Last_edited": "Last edited",
-	"Yours": "Yours"
+	"Yours": "Yours",
+	"EUDIW_Wallet_test": "EUDIW Wallet test"
 }
